@@ -1,37 +1,58 @@
 {
   "name": "portfolio",
   "version": "0.0.1",
-  "description": "Mon portfolio",
-  "productName": "portfolio",
+  "description": "Erick Golos est un développeur avec une expertise Front-end et spécialisé sur le Framework Vue.js.",
+  "productName": "Erick Golos - Portfolio",
   "author": "Erick Golos <45265019+Lurrik@users.noreply.github.com>",
+  "type": "module",
   "private": true,
   "scripts": {
+    "dev": "npx quasar dev",
+    "pwa": "npx quasar dev -m pwa",
+    "build": "npx quasar build",
+    "build-pwa": "npx quasar build -m pwa",
     "lint": "eslint --ext .js,.ts,.vue ./",
     "format": "prettier --write \"**/*.{js,ts,vue,scss,html,md,json}\" --ignore-path .gitignore",
     "test": "echo \"No test specified\" && exit 0"
   },
   "dependencies": {
-    "vue-i18n": "^9.2.2",
-    "@quasar/extras": "^1.0.0",
-    "quasar": "^2.6.0",
-    "vue": "^3.0.0",
-    "vue-router": "^4.0.0"
+    "@duskmoon/vue3-typed-js": "0.0.4",
+    "@quasar/extras": "1.16.12",
+    "@vueuse/motion": "2.0.0-beta.12",
+    "@vueuse/core": "11.1.0",
+    "axios": "1.7.2",
+    "bottom-navigation-vue": "3.3.5",
+    "pinia": "2.0.11",
+    "quasar": "2.16.6",
+    "register-service-worker": "1.7.2",
+    "vue": "3.5.12",
+    "vue-i18n": "9.2.2",
+    "vue-router": "4.4.0"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.10.0",
-    "@typescript-eslint/parser": "^5.10.0",
-    "eslint": "^8.10.0",
-    "eslint-plugin-vue": "^9.0.0",
-    "eslint-config-prettier": "^8.1.0",
-    "prettier": "^2.5.1",
-    "@types/node": "^12.20.21",
-    "@intlify/vite-plugin-vue-i18n": "^3.3.1",
-    "@quasar/app-vite": "^1.0.0",
-    "autoprefixer": "^10.4.2",
-    "typescript": "^4.5.4"
+    "@intlify/vite-plugin-vue-i18n": "6.0.3",
+    "@quasar/app-vite": "2.0.0-beta.12",
+    "@types/node": "20.5.9",
+    "@typescript-eslint/eslint-plugin": "7.7.0",
+    "@typescript-eslint/parser": "7.7.0",
+    "autoprefixer": "10.4.2",
+    "eslint": "8.56.0",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-plugin-vue": "9.27.0",
+    "prettier": "3.3.3",
+    "typescript": "5.5.3",
+    "vite-plugin-sitemap": "0.6.2",
+    "vue-tsc": "2.0.1",
+    "workbox-build": "7.1.1",
+    "workbox-cacheable-response": "7.1.0",
+    "workbox-core": "7.1.0",
+    "workbox-expiration": "7.1.0",
+    "workbox-precaching": "7.1.0",
+    "workbox-routing": "7.1.0",
+    "workbox-strategies": "7.1.0"
   },
   "engines": {
-    "node": "^18 || ^16 || ^14.19",
+    "node": "^24 || ^22 || ^20 || ^18",
     "npm": ">= 6.13.4",
     "yarn": ">= 1.21.1"
   }
