@@ -15,8 +15,7 @@ module.exports = {
   env: {
     browser: true,
     es2021: true,
-    node: true,
-    'vue/setup-compiler-macros': true
+    node: true
   },
 
   // Rules order is important, please avoid shuffling them
